chore: rename files folder to fileItems, address feedback
