+++ packages/react-storage/src/components/StorageBrowser/createStorageBrowser/__tests__/createProvider.spec.tsx
@@ -5,7 +5,7 @@ import { ActionConfigsProvider, getActionConfigs } from '../../actions';
 import { ComponentsProvider } from '../../components';
 import { createConfigurationProvider } from '../../configuration';
 import { DisplayTextProvider } from '../../displayText';
-import { FileItemsProvider } from '../../files';
+import { FileItemsProvider } from '../../fileItems';
 import { LocationItemsProvider } from '../../locationItems';
 import { StoreProvider } from '../../store';
 import { ActionHandlersProvider, getActionHandlers } from '../../useAction';
@@ -17,7 +17,7 @@ import createProvider from '../createProvider';
 jest.mock('../../actions');
 jest.mock('../../components');
 jest.mock('../../displayText');
-jest.mock('../../files');
+jest.mock('../../fileItems');
 jest.mock('../../locationItems');
 jest.mock('../../configuration');
 jest.mock('../../store');
