+++ packages/react-storage/src/components/StorageBrowser/displayText/types.ts
@@ -11,7 +11,7 @@ import type {
   LocationPermissions,
 } from '../actions';
 import type { MessageType } from '../components';
-import type { FileItems } from '../files';
+import type { FileItems } from '../fileItems';
 import type { LocationState } from '../store';
 
 /**
