+++ packages/react-storage/src/components/StorageBrowser/views/LocationActionView/DeleteView/__tests__/useDeleteView.spec.ts
@@ -8,7 +8,7 @@ import { useAction } from '../../../../useAction';
 
 import { useDeleteView } from '../useDeleteView';
 
-jest.mock('../../../../files');
+jest.mock('../../../../fileItems');
 jest.mock('../../../../locationItems');
 jest.mock('../../../../store');
 jest.mock('../../../../useAction');
